==> Step3_Sub-Subclustering_Tcells.Rmd <==
---
title: "T cell subclustering"
author: "Rosa Hernansaiz-Ballesteros, Ph.D; Vincent Perez, Ph.D. | Tempus Labs Inc."
date: "`r format(Sys.time(), '%d %B %Y')`"
output: 
  html_document:
      self_contained: true
      thumbnails: false
      lightbox: true
      gallery: false
      theme: cosmo
      highlight: kate
      toc: true
      toc_depth: 3
      toc_float: yes
      code_folding: show
params:
  seurat_scc: "~/vincent_perez/data/seurat_13Jun22_tcells_subclust.rds"
  cells: ["T cells", "NKT", "Tgd", "NK cells", "ILC"] 
  cluster_resolution: [0.075, 0.1, 0.15, 0.2, 0.3, 0.35, 0.4, 0.5, 0.8]
  prefix: "Tcells"
  output_fig: "Subclustering/Tcells"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning=FALSE, message=FALSE)
```

# Purpose
The objective is perform a sub-subclustering on the T cells, to potentially discriminate between activated and exhausted T cells.

# Setup

### Load libraries

```{r lib}
library(Seurat)
library(data.table)
library(janitor)
library(limma)
library(ggplot2)
library(ComplexHeatmap)
library(microViz)
library(dplyr)
library(harmony)
library(openxlsx)
library(ProjecTILs)
```


### Plot functions and colors
```{r functions}
theme_sara <- function(){
  theme_bw(base_size=10)+
    theme(axis.text=element_text(color="black"),
          panel.background=element_rect(color="white"),
          strip.text = element_text(size=12),
          strip.background = element_rect(fill="white"))
}

theme_sara_90 <- function() {
  theme_bw(base_size=18)+
    theme(axis.text.x=element_text(angle=90,hjust = 1,vjust = 0.5),
          axis.text=element_text(color="black"),
          panel.background=element_rect(color="black"),
          strip.text = element_text(size=12),
          strip.background = element_rect(fill="white"))
}
```

```{r plot-colors}
Group2.cols <- rcartocolor::carto_pal(n=5, "Teal")
Group4.cols <- rcartocolor::carto_pal(n=5, "Emrld")
Group10.cols <- rcartocolor::carto_pal(n=5, "BurgYl")
Group8.cols <- rcartocolor::carto_pal(n=5, "Purp")


sample.cols <- c(Group2.cols, Group4.cols, Group8.cols, Group10.cols)
names(sample.cols) <- c(c("770128", "770163", "770226", "770227", "770264"),
                        c("770200", "770223", "770238","770271", "770279"),
                        c("770174", "770184", "770225","770216", "770267"),
                        c("770240", "770224", "770231", "770203", "770232"))


group.cols <- c(RT=Group2.cols[[1]],
                Control=Group4.cols[[5]],
                RT_plus_CpG=Group8.cols[[1]],
                CpG=Group10.cols[[4]])

group.cols2=c("Ctl"="#535353",
              "CpG"="#72FA79",
              "RT"="#7A81FF",
              "CpG+RT"="#FF7E79")

unikn_blueRed <- colorRampPalette(c("#324376","white","#771434"))(256)

black_blue_yel <- colorRampPalette(c("black", "#31446B", "#666870", "#D9C560", "#FFE945"))(256)

black_blue_yel2 <- colorRampPalette(c("black", "#31446B", "#666870", "#d3c064", "#FFE945"))(256)

grey_blue_yel <- colorRampPalette(c("grey17", "#31446B", "#666870", "#D9C560", "#FFE945"))(256)

grey_blue_yel2 <- colorRampPalette(c("grey17", "#31446B", "#666870", "#d3c064", "#FFE945"))(256)

ref.cols <- list()

ref.cols[["SingleR_ImmGen"]]  <- c(`B cells` = "#F0A0FF", 
                                   `Basophils` = "#9370DB",
                                    `Eosinophils`= "#B452CD",
                                   `DC` = "#9DCC00", 
                                   `Epithelial cells` = "#87CEFA",  
                                   `Endothelial cells` = "#6495ED",
                                   `Fibroblasts` = "navy",
                                   `ILC` = "#191919", 
                                   `Macrophages` = "#1C8356", 
                                   `Mast cells` = "#E9Debb", 
                                   `Monocytes` = "#81C57A", 
                                   `Microglia` = "#EEB4B4",
                                   `Neutrophils` = "#b10da1",
                                   `NK cells` = "#8F7C00", 
                                   `NKT` = "#FEAF16", 
                                   `Others` = "gray40", 
                                   `Stem cells` = "red", 
                                   `Stromal cells` = "#F08080",
                                   `T cells` = "#FE902EFF", 
                                   `Tgd` = "#FFEE33")

ref.cols[["sex"]] = c("F" = "#bada55",  "M" = "#7a8888")
ref.cols[["phase"]] = c("G1" = "#ebd2be",  "S" = "#51cdc4",  "G2M" = "#cc0000")

heatmap.colfun <- microViz::heat_palette(
                  palette = "Lisbon",
                  breaks = 100,
                  range = c(0,1),
                  sym = FALSE,
                  rev = FALSE)

```

#### Specific functions

```{r specific-funs}
metaBarplot <- function(x, seurat, meta.var, sr){
  as.data.table(seurat@meta.data) %>%
    dplyr::group_by(!!as.name(names(meta.var)[[x]]), !!as.name(sr)) %>%
    dplyr::summarise(N = n()) %>%
    dplyr::mutate(res = factor(!!as.name(sr), levels = gtools::mixedsort(unique(seurat@meta.data[[sr]])))) %>%
    dplyr::rename(var = 1) %>%
    ggplot(., aes(x=res, y=N, fill=var)) + 
    geom_col(position="fill", color="black") + 
    theme_sara() +
    scale_fill_manual("", values=meta.var[[x]][which(names(meta.var[[x]]) %in% seurat@meta.data[[names(meta.var)[[x]]]])]) +
    labs(y = "Proportion", x = "") +
    guides(fill = guide_legend(ncol = 2, nrow = 10)) +
    theme(legend.position="right")
}

metaBarplot_nFeat <- function(selected.res, seurat_scc){
  cols_cluster <- pals::polychrome(n=seurat_scc[[paste0("RNA_snn_res.",selected.res)]] %>% pull() %>% levels() %>% length())
  names(cols_cluster) <- seurat_scc[[paste0("RNA_snn_res.",selected.res)]] %>% pull() %>% levels()
  
  as.data.table(seurat_scc@meta.data) %>%
    dplyr::select(!!as.name(paste0("RNA_snn_res.",selected.res)),nFeature_RNA, nCount_RNA, percent.mt) %>%
    tidyr::pivot_longer(cols = !paste0("RNA_snn_res.",selected.res), names_to = "Feature", values_to = "value") %>%
    dplyr::rename(res = 1) %>%
    dplyr::mutate(res = factor(res, levels = gtools::mixedsort(unique(seurat@meta.data[[paste0("RNA_snn_res.",selected.res)]])))) %>%
    ggplot(., aes(x=res, y=value, fill=res)) +
    geom_boxplot() +
    scale_fill_manual("", values = cols_cluster) +
    facet_wrap(.~Feature, scales = "free") +
    theme_sara() +
    ylab("") + xlab("") +
    guides(fill = guide_legend(ncol = 2, nrow = 6))
}

umap_dot_cluster <- function(x, seurat_scc, mm){
  print(x)
  cols_cluster <- pals::polychrome(n=seurat_scc[[x]] %>% pull() %>% levels() %>% length())
  names(cols_cluster) <- seurat_scc[[x]] %>% pull() %>% levels()
  Idents(seurat_scc) <- seurat_scc@meta.data[[x]]

  # UMAP
  p1 <- DimPlot(seurat_scc, group.by=x) + 
    theme_sara() +
    scale_color_manual(values=cols_cluster)
  
  # Dotplot
  res <- Map(`[`,  mm$cell_type_specific, relist(!duplicated(unlist( mm$cell_type_specific)), skeleton =  mm$cell_type_specific))
  
  seurat_scc@meta.data[[x]] = factor(seurat_scc@meta.data[[x]], 
                                     levels = gtools::mixedsort(unique(seurat_scc@meta.data[[x]])))
  
  p2 <- DotPlot(seurat_scc, features = res, scale=FALSE) + 
    theme_sara_90() + 
    geom_point(aes(size=pct.exp), shape = 21, colour="black", stroke=1) + 
    scale_color_gradientn(colors=grey_blue_yel, oob=scales::squish, na.value=0)

  return(list("umap" = p1, 
              "dotPlot" = p2))
}

```

```{r}

vln_plot <- function(seurat, cols, feature, log=TRUE, idents=NULL, alpha=0.5) {
  p <- Seurat::VlnPlot(seurat, idents=idents, features=feature, log=log, alpha=alpha)
  p <- p + theme_sara()
  p <- p + scale_fill_manual(values=cols)
  p <- p + theme(legend.position="none")
  return(p)
}

vln_box_plot <- function(dat, feature, cols, x) {
  p <- ggplot(data=dat, aes(x=.data[[x]], y=.data[[feature]], fill=.data[[x]]))
  p <- p + geom_boxplot(width=0.5, alpha=0.6)
  p <- p + geom_violin(alpha=0.5, width=1, position="identity") + labs(y= paste0("log2 (",feature, ")"))
  p <- p + theme_sara() + theme(axis.title.x=element_blank())
  p <- p + scale_fill_manual(values=cols)
  return(p)
}

features_boxplot <- function(seurat, res, colors, subclusters = NULL, features){
  
  Idents(seurat) <- seurat@meta.data[[res]]
  
  if(is.null(subclusters)){
    all <- as.data.frame(t(log2(seurat[["RNA"]]@data[features,]+1) %>% as.matrix())) %>%
      tibble::rownames_to_column(var = "rn")

  }else{
    specific_cells <- WhichCells(seurat, idents=as.character(subclusters))
    all <- as.data.frame(t(log2(seurat[["RNA"]]@data[features, specific_cells]+1) %>%  as.matrix())) %>%
      tibble::rownames_to_column(var = "rn")
  }

  dat <- as.data.frame(seurat@meta.data) %>%
    tibble::rownames_to_column(var = "rn")
  all <- merge(dat, all, by="rn", all.y=TRUE) %>%
    dplyr::mutate(res = factor(!!as.name(res), 
                               levels=gtools::mixedsort(unique(seurat@meta.data[[res]]))))

  ps = lapply(features, function(x, all, colors, res){
    vln_box_plot(dat=all, feature=x, cols=colors, x = res) +
      guides(fill = FALSE) #guide_legend(ncol = 2)
  }, all, colors, res)
  
  names(ps) = features
  return(ps)

}

```

### Import Data
```{r}
seurat_scc <- readRDS(params$seurat)

cols_01 <- pals::polychrome(n=seurat_scc[["RNA_snn_res.0.3"]] %>% pull() %>% levels() %>% length())
names(cols_01) <- seurat_scc[["RNA_snn_res.0.3"]] %>% pull() %>% levels()
cols_cluster <- c(cols_01["0"], "1_0" = "#E4E1E3", "1_1" = "#F3E5AB", cols_01[as.character(2:9)]) 

seurat_harmony = readRDS("~/vincent_perez/data/seurat_30May22_harmony_clusters.rds")

```

```{r load-markers}
mmarkers <- readr::read_csv("~/vincent_perez/data/Murine_SingleCell_Immunology_Markers - Markers.csv")
names(mmarkers) <- janitor::make_clean_names(names(mmarkers))
mmarkers_scc <- mmarkers %>%
  dplyr::filter(cell_type_broad %in% c("T cell", "ILC")) %>%
  dplyr::filter(!is.na(mm10_gene))

mm <- list()
mm[["cell_type_broad"]] = lapply(unique(mmarkers_scc$cell_type_broad), function(x){
  features = mmarkers_scc %>%
    dplyr::filter(mm10_gene %in% rownames(seurat_harmony[["RNA"]]@data)) %>%
    dplyr::filter(cell_type_broad == x) %>%
    dplyr::pull(mm10_gene) %>% unique()
  
})
names(mm[["cell_type_broad"]]) = unique(mmarkers_scc$cell_type_broad)

mm[["cell_type_specific"]] = lapply(unique(mmarkers_scc$cell_type_specific), function(x){
  features = mmarkers_scc %>%
    dplyr::filter(mm10_gene %in% rownames(seurat_harmony[["RNA"]]@data)) %>%
    dplyr::filter(cell_type_specific == x) %>%
    dplyr::pull(mm10_gene) %>% unique()
  
})
names(mm[["cell_type_specific"]]) = unique(mmarkers_scc$cell_type_specific)
```
# Positive CD8 Tcells

# Sample distribution in detail

## Per sample

Overview on the proportion of cells per sample. The proportion is calculated based on the total number of cells per sample.

```{r sample-overview}

sfull = as.data.table(seurat_harmony@meta.data) %>%
  dplyr::group_by(SampleID) %>%
  dplyr::summarise(n_cell=n())

summary.df <- as.data.table(seurat_scc@meta.data) %>% 
  dplyr::group_by(SampleID, SingleR_ImmGen) %>% 
  dplyr::summarise(n=n()) %>%
  dplyr::left_join(., unique.data.frame(as.data.table(seurat_scc@meta.data[, c("SampleID", "Group")])), by = "SampleID") %>%
  dplyr::left_join(., sfull, by = "SampleID") %>%
  dplyr::mutate(p = n/n_cell)

p2 <- ggplot(data=summary.df, aes(x=reorder(SampleID, -n), y=p, fill=SingleR_ImmGen))
p2 <- p2 + geom_bar(stat="identity", color="black")
p2 <- p2 + scale_fill_manual('', values=ref.cols$SingleR_ImmGen[which(names(ref.cols$SingleR_ImmGen) %in% summary.df$SingleR_ImmGen)])
p2 <- p2 + theme_sara_90() 
p2 <- p2 + theme(axis.text=element_text(size=9), 
                 axis.title.x = element_blank(), 
                 axis.title.y=element_text(size=11))
p2 <- p2 + labs(y="Proportion of cells") 
p2 <- p2 + facet_wrap(~Group, scale="free_x")
p2

# ggsave(file.path(params$output_fig, paste0(params$prefix, "_proportion_cells_sample.png")), 
#        width=6, height=5,plot = p2, units="in", dpi=400)
```

```{r sample-overview-2}

summary.df %>%
    ggplot(., aes(x=reorder(SampleID, -n), y=n, fill=SingleR_ImmGen)) + 
    geom_col(position="fill", color="black") + 
    theme_sara_90() +
    scale_fill_manual('', values=ref.cols$SingleR_ImmGen[which(names(ref.cols$SingleR_ImmGen) %in% summary.df$SingleR_ImmGen)]) +
    labs(y = "Proportion", x = "") +
    facet_wrap(~Group, scale="free_x") +
    guides(fill = guide_legend(ncol = 2, nrow = 10)) +
    theme(legend.position="right")

# ggsave(file.path(params$output_fig, paste0(params$prefix, "_proportion_cells_sample-2.png")), 
#        width=7, height=6, units="in", dpi=400)
```

```{r sample-overview-3}

summary.df <- as.data.table(seurat_scc@meta.data) %>% 
  dplyr::group_by(RNA_snn_res.0.3_sub1.0.1, SampleID, SingleR_ImmGen) %>% 
  dplyr::summarise(n=n()) %>%
  dplyr::left_join(., unique.data.frame(as.data.table(seurat_scc@meta.data[, c("SampleID", "Group")])), by = "SampleID")

summary.df %>%
    ggplot(., aes(x=reorder(SampleID, -n), y=n, fill=SingleR_ImmGen)) + 
    geom_col(position="fill", color="black") + 
    theme_sara_90() +
    scale_fill_manual('', values=ref.cols$SingleR_ImmGen[which(names(ref.cols$SingleR_ImmGen) %in% summary.df$SingleR_ImmGen)]) +
    labs(y = "Proportion", x = "") +
    facet_grid(RNA_snn_res.0.3_sub1.0.1~Group, space="free", scales= "free") +
    guides(fill = guide_legend(ncol = 5, nrow = 1)) +
    theme(legend.position="bottom")

# ggsave(file.path(params$output_fig, paste0(params$prefix, "_proportion_cells_sample-3.png")), 
#        width=12, height=14, units="in", dpi=400)
```

## Per group
Creating barplots for the whole group

```{r}

sfull = as.data.table(seurat_harmony@meta.data) %>%
  dplyr::group_by(Group) %>%
  dplyr::summarise(n_cell=n())

summary.df <- as.data.table(seurat_scc@meta.data) %>%
  dplyr::group_by(Group, SingleR_ImmGen) %>%
  dplyr::summarise(n=n()) %>% 
  dplyr::left_join(., sfull, by = "Group") %>%
  dplyr::mutate(p = n/n_cell)

p2 <- ggplot(data=summary.df, aes(x=Group, y=p, fill=SingleR_ImmGen))
p2 <- p2 + geom_bar(stat="identity", color="black")
p2 <- p2 + scale_fill_manual(values=ref.cols$SingleR_ImmGen[which(names(ref.cols$SingleR_ImmGen) %in% summary.df$SingleR_ImmGen)])
p2 <- p2 + theme_sara_90() 
p2 <- p2 + theme(axis.text=element_text(size=9), 
                 axis.title.x = element_blank(), 
                 axis.title.y=element_text(size=11),
                 legend.title = element_blank(),
                 legend.position="bottom")
p2 <- p2 + labs(y="Proportion of cells") 
p2

# ggsave(file.path(params$output_fig, paste0(params$prefix, "_proportion_cells_group.png")), 
#        width=6, height=5, units="in", dpi=400)

```

# Subclustering for Clusters 2, 3 and 5

## Current clustering composition

Current cluster was obtained using a general resolution of 0.3.
Cluster 1 was then sub-clustered using a resolution of 0.1, 
so the different naive populations were discriminated.

```{r}
seurat_scc@meta.data$RNA_snn_res.0.3_sub1.0.1_V2<-seurat_scc@meta.data$RNA_snn_res.0.35
# seurat_scc@meta.data$RNA_snn_res.0.3_sub1.0.1_V2[which(seurat_scc@meta.data$RNA_snn_res.0.35 %in% c("1_0","1_1"))]<-"1"

current.res =  "RNA_snn_res.0.3_sub1.0.1_V2"

cols_cluster2<-c("0"="#5A5156",
                 "1"="#E4E1E3",
                 "2"="#F6222E",
                 "3"="#FE00FA",
                 "4"="#16FF32",
                 "5"="#3283FE",
                 "6"="#FEAF16",
                 "7"="#B00068",
                 "8"="#1CFFCE",
                 "9"="#90AD1C")

p<-DimPlot(seurat_scc, group.by=current.res) + 
    theme_sara() #+
#    scale_color_manual(values=cols_cluster2)
p

# pdf(file = "~/vincent_perez/Chang_T_cell_clusters_UMAP.pdf",
#     width = 8,
#     height = 6
#     )
# p
# dev.off

# ggsave(file.path(params$output_fig,"UMAP_clusters_res0.3_sub1.0.1.png"), height=4, width=7, units="in", device="png", dpi=400)


metaBarplot(1, seurat_scc, list("SingleR_ImmGen" = ref.cols$SingleR_ImmGen), current.res)

# ggsave(file.path(params$output_fig,"Barplot_clusters_res0.3_sub1.0.1.png"), height=4, width=5, units="in", device="png", dpi=400)


### Plot treatment
seurat_scc@meta.data$Group2<-seurat_scc@meta.data$Group
seurat_scc@meta.data$Group2[which(seurat_scc@meta.data$Group2 %in% "Control")]<-"Ctl"
seurat_scc@meta.data$Group2[which(seurat_scc@meta.data$Group2 %in% "RT_plus_CpG")]<-"CpG+RT"
seurat_scc@meta.data$Group2<-factor(seurat_scc@meta.data$Group2, levels=c("Ctl", "CpG", "RT","CpG+RT"))
# group.cols2=c("Ctl"="#3cb44b",
#            "CpG"="navy",
#            "RT"="#f032e6",
#            "CpG+RT"="orange")

p<-DimPlot(seurat_scc, 
           group.by="Group2") + 
    theme_sara() +
    scale_color_manual(values=group.cols2) +
  ggtitle("Treatment group")
p

# pdf(file = "~/vincent_perez/Chang_T_cell_treatment_UMAP.pdf",
#     width = 8,
#     height = 6
#     )
# p
# dev.off

### Plot cell cycle
phase.cols<-c("G1"="#EBD2BE",
              "G2M"="#CC0600",
              "S"="#51CDC4")

p<-DimPlot(seurat_scc, 
           group.by="Phase") + 
    theme_sara() +
    scale_color_manual(values=phase.cols)
p

# pdf(file = "~/vincent_perez/Chang_T_cell_cellcycle_UMAP.pdf",
#     width = 8,
#     height = 6
#     )
# p
# dev.off

### Barplot for phase
p<-metaBarplot(1, seurat_scc, list("Phase" = phase.cols), current.res)
p

# pdf(file = "~/vincent_perez/Chang_T_cell_cellcycle_barplot.pdf",
#     width = 6,
#     height = 5
#     )
# p
# dev.off()


### Barplot for phase, alternative
df<-as.data.table(seurat_scc@meta.data) %>%
  group_by(RNA_snn_res.0.3_sub1.0.1_V2,Phase) %>%
  summarise(n = n()) %>%
  mutate(freq=n/sum(n)) 

lvls<-df %>% filter(Phase=="G1") %>%
  arrange(freq)

df$RNA_snn_res.0.3_sub1.0.1_V2<-factor(df$RNA_snn_res.0.3_sub1.0.1_V2, levels=c(3,5,11,6,12,10,2,4,9,1,8,7,0))

p<-ggplot(df, aes(x=RNA_snn_res.0.3_sub1.0.1_V2, y=freq, fill=Phase)) + 
    geom_col(position="fill", color="black") + 
  theme_sara() +
  scale_fill_manual(values=phase.cols) +
    labs(y = "Proportion", x = "") +
    guides(fill = guide_legend(ncol = 2, nrow = 10)) +
    theme(legend.position="right") +
  geom_hline(yintercept = 0.5, linetype = "dotted", color = "darkgrey")
p

pdf(file = "~/vincent_perez/figures/Chang_T_cell_cellcycle_barplot2.pdf",
    width = 6,
    height = 5
    )
p
dev.off()

### And for treatment
group.cols3=c("Ctl"="#535353",
           "CpG"="#72FA79",
           "RT"="#7A81FF",
           "CpG+RT"="#FF7E79")
p<-metaBarplot(1, seurat_scc, list("Group2" = group.cols3), current.res)


p

# pdf(file = "~/vincent_perez/Chang_T_cell_treatment_barplot.pdf",
#     width = 6,
#     height = 5
#     )
# p
# dev.off()


#### Flip the barplot
cols_cluster3=c("0"="#5A5156",
               "1"="#E4E1E3",
               "2"="#2ED9FF",
               "3"="#1CFFCE",
               "4"="#3283FE",
               "5"="#FE00FA",
               "6"="#90AD1C",
               "7"="#16FF32",
               "8"="#B00068",
               "9"="#FEAF16",
               "10"="#F6222E",
               "11"="#AA0DFE",
               "12"="#DEA0FD")
p<-metaBarplot(1, seurat_scc, list("RNA_snn_res.0.3_sub1.0.1_V2" = cols_cluster3), "Group2")


p

# pdf(file = "~/vincent_perez/figures/Chang_T_cell_cluster_barplot.pdf",
#     width = 6,
#     height = 5
#     )
# p
# dev.off()


```

## Plot GOIs

```{r, , fig.height=6, fig.width=6}



gois<-c('Ncr1', ### Cluster 0,
        'Ccl5','Gzma','Gzmb', 'Ly6c2', 'Cxcr6', 'Cd28', 'Gzmk', 'Tnfaip3', 'Tnfaip8', ### Cluster 1
        'Pdcd1', 'Ctla4',  'Havcr2', 'Tox', ### Cluster 1
        'CD4', 'Lef1', 'S1pr1', 'Ccr7', ### Cluster 2
        'Mki67', 'Top2a', 'Birc5', ### Cluster 3
        'Runx3', 'Pdcd1', 'Havcr2', ### Cluster 3
        'Foxp3', ### Cluster 4
        'Runx3', 'Cdca7',  'Stmn1', 'Txn1',  #'Gzmk', 'Gzmb', 'Gzma','Gzmm','Gzme',"Gzmd","Gzmg","Gzmn","Gzmf","Gzmc", ### Cluster 5
        'Pdcd1',  'Tox', 'Eomes', ### Cluster 5
        'Lef1', 'Ccr7','Il7r', ### Cluster 6
        'Ly6c2', ### Cluster 7
        'Ncr1', ### Cluster 8
        'Tcrg-C1', ### Cluster 9
        'Tbc1d4','Tnfsf8', ### Cluster 10 
        'Slamf6', 'Eomes', 'Foxp1', ### Cluster 11
        'Cd74' , 'H2-Aa',  ### Cluster 12
        'Gzmk', 'Ifng', 'Lag3', 'Havcr2', 'Tox', 'Cd4','Cd8a'  ### Other
        )
gois<-unique(gois)
gois<-rev(gois)

### Set order
seurat_scc@meta.data$RNA_snn_res.0.3_sub1.0.1_V2<-factor(seurat_scc@meta.data$RNA_snn_res.0.3_sub1.0.1_V2, levels=c('0', '8',  '9', '10', '4','2','1', '3', '5', '6', '7',  '11', '12'))

Idents(seurat_scc)<-seurat_scc@meta.data$RNA_snn_res.0.3_sub1.0.1_V2
dot<-DotPlot(seurat_scc, features = gois) +theme_sara_90() + coord_flip()
dot

pdf(file = "~/vincent_perez/figures/Chang_T_cell_dotplot.pdf",
    width = 7,
    height = 8
    )
dot
dev.off()
```

## Plot GOIs (alternative)

```{r, fig.height=6, fig.width=6}
markers<-read.csv("~/vincent_perez/data/de_tcells_res0.35_kim.csv")
top<-markers %>% 
    group_by(cluster) %>%
    top_n(n=3, wt=avg_log2FC)

# gois2<-gois[which(gois %in% top$gene)]
# dot2<-DotPlot(seurat_scc, features = gois2) +theme_sara_90() + coord_flip()
# dot2

### Place c('cd4', 'cd8a', 'lag3', 'havcr2', 'tox','pdcd1','ctla4','Txn1','Eomes','Gzmk', 'Gzmb', 'Gzma','Gzmm','Gzme',"Gzmd","Gzmg","Gzmn","Gzmf","Gzmc", 'Ifng') in a separate dotplot
top<-c(top$gene, gois)
index<-which(top %in% c("Tnfaip3", "Txn1"))
top2<-top[-index]
top3<-rev(unique(top2))


### Set order
top3_gois<-c(# "Ncr1", "Cxcr6", "Cd28", "Tnfaip8", "S1pr1", "Top2a", "Birc5", "Runx3",  "Foxp3", "Cdca7", "Stmn1", "Il7r", 
        "H2-Aa", "Cd74", ### Cluster 12
        "Tspan3", "Retreg1", "Myb", "Slamf6",  ### Cluster 11
        "Tnfsf8", "Tbc1d4", 'Tnfaip3',### Cluster 10
        "Il17a", "Tcrg-C1", "Trdc", 'Il7r',### Cluster 9
        "Gzmb","Serpinb9b", "Gzma", ### Cluster 8
        "Ifit3", "Isg15", "Ifit1", ### Cluster 7
        "Ccr7", "Dapl1", ### Cluster 6
        "Mcm3", "Dut",'Cdca7','Stmn1', ### Cluster 5
        "Ctla4", "Tnfrsf4", "Ikzf2", 'Foxp3',### Cluster 4
        "Mki67","Hist1h2ap", "Hist1h1b",'Top2a','Birc5','Runx3', ### Cluster 3
        "Dusp10", "Rgs10", "Lef1", 'S1pr1',### Cluster 2
        "Cd8b1", "Ly6c2", "Ccl5",'Cxcr6','Cd28', ### Cluster 1
        "Spp1", "Tyrobp", "Gzmc",'Ncr1')### Cluster 0

dot3<-DotPlot(seurat_scc, features = top3_gois) +theme_sara_90() + coord_flip()
dot3

pdf(file = "~/vincent_perez/figures/Chang_T_cell_dotplot2.pdf",
    width = 7,
    height = 8.1
    )
dot3
dev.off()
```


## Plot expansion/exhaustion GOIs

```{r, fig.height=5, fig.width=5}
### Place in a separate dotplot
ee_gois<- c('Gzmk', 
            'Gzmb', 
            'Gzma',
#            'Gzmm',
#            'Gzme',
#            "Gzmd",
#            "Gzmg",
#            "Gzmn",
#            "Gzmf",
#            "Gzmc"
            'Lag3', 'Havcr2', 
            'Tox','Pdcd1',
            'Ctla4','Txn1',
            'Ifng',
            'Eomes',
            'Cd4', 'Cd8a' 
) 

dot4<-DotPlot(seurat_scc, features = ee_gois) +theme_sara_90() + coord_flip()
dot4

pdf(file = "~/vincent_perez/figures/Chang_T_cell_activation_markers_dotplot.pdf",
    width = 7,
    height = 5
    )
dot4
dev.off()
```

## Subclustering 

We want to discriminate between the active and exhausted cell from clusters 2, 3 and 5.
So we will sub-cluster these groups to try to find these groups.

```{r subclustering-1}
de.res = "RNA_snn_res.0.3_sub1.0.1"

res_sub = c("2"=0.2,"3"=0.1,"5"=0.3)

for(i in 1:length(res_sub)){
  if(i == 1){
    Idents(seurat_scc) <- seurat_scc@meta.data[[de.res]]
    subcluster.name = paste0("RNA_snn_res.0.3_sub")
    seurat_sub = Seurat::FindSubCluster(seurat_scc, 
                                          cluster = as.numeric(names(res_sub)[i]), 
                                          graph.name = "RNA_snn", 
                                          resolution = res_sub[i], 
                                          algorithm = 2,
                                          subcluster.name = subcluster.name)

  }else{
    print(i)
    Idents(seurat_sub) <- seurat_sub@meta.data[[subcluster.name]]
    seurat_sub = Seurat::FindSubCluster(seurat_sub, 
                                          cluster = as.numeric(names(res_sub)[i]), 
                                          graph.name = "RNA_snn", 
                                          resolution = res_sub[i], 
                                          algorithm = 2,
                                          subcluster.name = subcluster.name)
  }

}

# saveRDS(seurat_sub, "seurat_28Jun22_tcells_subclustr1235.rds")

```

## Analysis of subclusters

```{r}

cols_subcluster = c(cols_cluster[1:3], 
                    "2_0" = "#F6222E", "2_1" = "#F8646C", "2_2" = "#7B1117",
                    "3_0" = "#FE00FA", "3_1" = "#FFA7CE",
                    cols_cluster[6],
                    "5_0" = "#3283FE", "5_1" = "#ACE5EE",
                    cols_cluster[8:11])

subcluster.name = "RNA_snn_res.0.3_sub"

DimPlot(seurat_sub, group.by=subcluster.name) + 
    theme_sara() +
    scale_color_manual(values=cols_subcluster)

# ggsave(file.path(params$output_fig,"UMAP_clusters_res0.3_sub1235.png"), height=4, width=7, units="in", device="png", dpi=400)


metaBarplot(1, seurat_sub, list("SingleR_ImmGen" = ref.cols$SingleR_ImmGen), subcluster.name)

# ggsave(file.path(params$output_fig,"Barplot_clusters_res0.3_sub1235.png"), height=4, width=6, units="in", device="png", dpi=400)


cluster_fig <- umap_dot_cluster(subcluster.name, seurat_sub, mm)
cluster_fig$dotPlot

```

```{r sample-overview-4}

summary.df <- as.data.table(seurat_sub@meta.data) %>% 
  dplyr::group_by(RNA_snn_res.0.3_sub, SampleID, SingleR_ImmGen) %>%
  dplyr::filter(grepl("_", RNA_snn_res.0.3_sub, fixed = T)) %>%
  dplyr::filter(!grepl("1_", RNA_snn_res.0.3_sub, fixed = T)) %>%
  dplyr::summarise(n=n()) %>%
  dplyr::left_join(., unique.data.frame(as.data.table(seurat_sub@meta.data[, c("SampleID", "Group")])), by = "SampleID")

summary.df %>%
    ggplot(., aes(x=reorder(SampleID, -n), y=n, fill=SingleR_ImmGen)) + 
    geom_col(position="fill", color="black") + 
    theme_sara_90() +
    scale_fill_manual('', values=ref.cols$SingleR_ImmGen[which(names(ref.cols$SingleR_ImmGen) %in% summary.df$SingleR_ImmGen)]) +
    labs(y = "Proportion", x = "") +
    facet_grid(RNA_snn_res.0.3_sub~Group, space="free", scales= "free") +
    guides(fill = guide_legend(ncol = 5, nrow = 1)) +
    theme(legend.position="bottom")

# ggsave(file.path(params$output_fig, paste0(params$prefix, "_proportion_cells_sub1235_sample-3.png")), 
#        width=12, height=10, units="in", dpi=400)
```

## DE (Find markers)

```{r}
# markers.list <- list()
# top10.list <- list()
# 
# de.res = "RNA_snn_res.0.3_sub"
# 
# for(i in de.res){
#   Idents(seurat_sub) <- seurat_sub@meta.data[[i]]
#   markers.list[[i]] <- FindAllMarkers(seurat_sub, 
#                                                  only.pos = TRUE,
#                                                  min.pct = 0.25, 
#                                                  logfc.threshold = .1,
#                                                  random.seed = 888)
# 
# markers.list[[i]] %>%
#   dplyr::group_by(cluster) %>%
#   dplyr::top_n(n = 10, wt = avg_log2FC) -> top10.list[[i]]
# 
# }

```

Export
```{r}
# results <- list(top10=top10.list[[de.res]],
#                 top30=markers.list[[de.res]] %>%
#                                dplyr::group_by(cluster) %>%
#                                dplyr::top_n(n = 30, wt = avg_log2FC))
# 
# header_style <- createStyle(halign = "center", textDecoration = "bold")
# 
# wb <- createWorkbook()
# 
# addWorksheet(wb, "Top10")
# writeData(wb, "Top10", results$top10, headerStyle = header_style)
# freezePane(wb, "Top10", firstRow = TRUE)
# setColWidths(wb, "Top10", cols = 1:ncol(results$top10), widths = "auto")
# 
# addWorksheet(wb, "Top30")
# writeData(wb, "Top30", results$top30, headerStyle = header_style)
# freezePane(wb, "Top30", firstRow = TRUE)
# setColWidths(wb, "Top30", cols = 1:ncol(results$top30), widths = "auto")
# 
# addWorksheet(wb, "All")
# writeData(wb, "All", markers.list[[de.res]], headerStyle = header_style)
# freezePane(wb, "All", firstRow = TRUE)
# setColWidths(wb, "All", cols = 1:ncol(markers.list[[de.res]]), widths = "auto")
# 
# 
# # saveWorkbook(wb, file = "Tcells_Louvain_subcluster1235_top_markers.xlsx", overwrite = TRUE)

```

### Heatmap
```{r heatmap-markers}
# heatmap.res = "RNA_snn_res.0.3_sub"
# 
# markers.list[[heatmap.res]] %>%
#   dplyr::group_by(cluster) %>%
#   dplyr::top_n(n = 6, wt = avg_log2FC) -> top6
# 
# 
# marker.genes <- top6$gene
# SetIdent(seurat_scc, value=heatmap.res)
# seurat.hm <- subset(seurat_scc, downsample=200)
# mat<- seurat.hm[["RNA"]]@data[marker.genes, ] %>% as.matrix()
# mat<- log2(mat+1)
```

```{r, fig.height=12, fig.width=10}
# clust_anno_top <- seurat.hm@meta.data[[heatmap.res]]
# clust_anno_row <- as.character(top6$cluster)
# 
# ra <- ComplexHeatmap::rowAnnotation(`Cluster`=clust_anno_row,
#                                      col=list(`Cluster`=cols_subcluster),
#                                      show_legend=FALSE)
# 
# ta <- ComplexHeatmap::HeatmapAnnotation(`Cluster`=clust_anno_top,
#                                         `Phase` = seurat.hm@meta.data$Phase,
#                                         `Sex` = seurat.hm@meta.data$sex,
#                                         `Group` = seurat.hm@meta.data$Group,
#                                         `Cell Type` = seurat.hm@meta.data$SingleR_ImmGen,
#                                      col=list(`Cluster`=cols_subcluster,
#                                               `Phase` = ref.cols$phase,
#                                               `Sex` = ref.cols$sex,
#                                               `Group` = group.cols,
#                                               `Cell Type` = ref.cols$SingleR_ImmGen),
#                                      show_legend=TRUE)
# 
# ComplexHeatmap::Heatmap(mat, name = "Expression",  
#                         cluster_columns = FALSE,
#                         cluster_rows=FALSE,
#                         column_split=seurat.hm@meta.data[[heatmap.res]],
#                         row_split=factor(top6$cluster, levels = gtools::mixedsort(as.character(unique(top6$cluster)))),
#                         row_names_gp = grid::gpar(fontsize = 8),
#                         column_title=character(0),
#                         column_gap = unit(0.5, "mm"),
#                         col = grey_blue_yel2,
#                         top_annotation = ta,
#                         show_column_names = FALSE,
#                         left_annotation = ra)
# 
# # png(file.path(params$output_fig,"Heatmap_Louvain_0.3_sub1235.png"), width=12, height=12, units="in", res=400)
# ComplexHeatmap::Heatmap(mat, name = "Expression",  
#                         cluster_columns = FALSE,
#                         cluster_rows=FALSE,
#                         column_split=seurat.hm@meta.data[[heatmap.res]],
#                         row_split=factor(top6$cluster, levels = gtools::mixedsort(as.character(unique(top6$cluster)))),
#                         row_names_gp = grid::gpar(fontsize = 8),
#                         column_title=character(0),
#                         column_gap = unit(0.5, "mm"),
#                         col = grey_blue_yel2,
#                         top_annotation = ta,
#                         show_column_names = FALSE,
#                         left_annotation = ra)
# # dev.off()
```

## Markers of interest

```{r}
sub_mm = list(`CD8+` = c("Cd3e","Cd8a","Havcr2","Lag3","Pdcd1","Cd44"),
              `Activated CD8+` = c("Tcf7","Gzmb","Ifng","Tnf","Il2"),
              `Exhausted CD8+` = c("Ctla4","Btla","Tigit","Eomes","Cd69"))

Idents(seurat_sub) <- seurat_sub@meta.data[[subcluster.name]]

#   seurat_scc@meta.data[[x]] = factor(seurat_scc@meta.data[[x]], 
#                                      levels = gtools::mixedsort(unique(seurat_scc@meta.data[[x]])))
#   
# ggdat = FetchData(seurat_sub, 
#                   var = c(union(mm$cell_type_specific$`Activated CD8+ T cell`, mm$cell_type_specific$`Exhausted CD8+ T cell`), subcluster.name)) %>%
#   as.data.frame() %>%
#   dplyr::rename(res = 4) %>%
#   # dplyr::filter(Tlr9 > 0 ) %>%
#   dplyr::mutate(Tlr9 = Tlr9+1)

DotPlot(seurat_sub, 
        features = sub_mm, 
        scale=FALSE, 
        idents = c("2_0", "2_1", "2_2", "3_0", "3_1", "5_0", "5_1"),
        cluster.idents = T) + 
    theme_sara_90() + 
    geom_point(aes(size=pct.exp), shape = 21, colour="black", stroke=1) + 
    scale_color_gradientn(colors=grey_blue_yel, oob=scales::squish, na.value=0)

# ggsave(file.path(params$output_fig, paste0(params$prefix, "_markers_interest_sub1235.png")), 
#        width=10, height=5,units="in", dpi=400)

```

```{r}
# 
# FetchData(seurat_sub,
#                   var = c("Cd4", "Cd8a", subcluster.name, "SampleID", "Group", "SingleR_ImmGen")) %>%
#   as.data.frame() %>%
#   rename(res =3) %>%
#   dplyr::mutate(size = if_else(res == "2_2", 0.65, 0.5)) %>%
#   ggplot(., aes(Cd4, Cd8a, color = res)) +
#   geom_point(aes(size = size)) +
#   theme_sara() +
#   scale_color_manual("",values = cols_subcluster) +
#   guides(size = FALSE)


```

## Violin - major groups

```{r}
featrues <- list(major = c("Cd4", "Cd8a", "Foxp3", "Ncr1", "Tcrg-C1"),
                 cd4 = c("Cd4", "Foxp3","Tigit", "Lef1", "Maf", "Tox", "Tbc1d4", "Icos", "Cxcr5"),
                 cd8 = c("Cd8a", "Ifng", "Ccr7", "Pdcd1", "Tcf7", "Gzmb", "Tnf" , "Il2", "Eomes", "Ctla4"))
```


```{r}
major_plots = features_boxplot(seurat_sub, res = subcluster.name, 
                               colors = cols_subcluster, 
                               features = featrues$major)

p <- cowplot::plot_grid(major_plots$Cd4,
                    major_plots$Cd8a,
                    major_plots$Ncr1,
                    major_plots$`Tcrg-C1`,
                   adjust="vh",
                   nrow=2,
                   ncol=2,
                   labels="AUTO")

# ggsave(file.path(params$output_fig,"Tcells_violin_major_sub1235.png"), height=8, width=10, units="in",
#        plot=p, device="png", dpi=400)
p
```

```{r}
subcltr = list(cd4 = c("1_0","2_2", 4,7), cd8 = c("1_1", "2_0", "2_1", "2_2", "3_0", "3_1", "5_0", "5_1", 9))
```

## Violin - Cd4
```{r}

cd4_plots = features_boxplot(seurat_sub, res = subcluster.name, 
                             colors = cols_cluster, subclusters = subcltr$cd4,
                             features = featrues$cd4)

p <- cowplot::plot_grid(cd4_plots$Foxp3,
                    cd4_plots$Lef1,
                    cd4_plots$Maf,
                    cd4_plots$Tox,
                    cd4_plots$Tbc1d4,
                    cd4_plots$Icos,
                    adjust="vh",
                    nrow=2,
                    ncol=3,
                    labels="AUTO")

# ggsave(file.path(params$output_fig,"Tcells_violin_CD4_sub1235.png"), height=8, width=10, units="in", device="png", dpi=400)
p

```

## Violin  - CD8
```{r}

cd8_plots = features_boxplot(seurat_sub, res = subcluster.name, 
                             colors = cols_subcluster, 
                             subclusters = subcltr$cd8,
                             features = featrues$cd8)

p <- cowplot::plot_grid(cd8_plots$Ccr7,
                    cd8_plots$Pdcd1,
                    cd8_plots$Eomes,
                    cd8_plots$Ctla4,
                    cd8_plots$Ifng,
                    cd8_plots$Tcf7,
                    cd8_plots$Gzmb,
                    cd8_plots$Tnf,
                    adjust="vh",
                    nrow=2,
                    ncol=4,
                    labels="AUTO")

# ggsave(file.path(params$output_fig,"Tcells_violin_CD8_sub1235.png"), 
#        height=6, width=10, units="in", device="png", dpi=400)

p
```

```{r}

# FetchData(seurat_scc,
#                   var = c( "Cd8a", "RNA_snn_res.0.3_sub1.0.1", "SampleID", "Group", "SingleR_ImmGen")) %>%
#   as.data.frame() %>%
#   rename(res = 2) %>% head()
```


# T cell types by group

```{r}
seurat_sub@meta.data$L5_clust <- as.factor(as.character(seurat_sub@meta.data[[subcluster.name]]))
seurat_sub@meta.data$Major_T_types <- forcats::fct_collapse(seurat_sub@meta.data$L5_clust,
                                                            CD4=subcltr$cd4,
                                                            CD8=subcltr$cd8,
                                                            ILC="0",
                                                            Tgd="6",
                                                            Other="8")
tcell.cols <- c(`CD4`="#F6222E",
                `CD8`="#3283FE",
                `ILC`="#5A5156",
                `Tgd`="#FFEE33",
                Other="grey90")

# saveRDS(seurat_sub, file="./../processedData/Subclustering/Tcells/seurat_tcells.rds")
```

```{r}
dt1 <- seurat_sub@meta.data
dt <- dt1 %>%
  dplyr::group_by(Group, Major_T_types) %>%
    dplyr::summarise(N = n()) %>%
    dplyr::mutate(Major_T_types = factor(Major_T_types, levels = gtools::mixedsort(unique(dt1$Major_T_types)))) 

dt$total_cells <- dt$Group
dt$total_cells <- gsub("RT$", "29159", dt$total_cells, perl=TRUE)
dt$total_cells <- gsub("RT_plus_CpG", "27056", dt$total_cells)
dt$total_cells <- gsub("Control", "19997", dt$total_cells)
dt$total_cells <- gsub("CpG", "17881", dt$total_cells)
dt$total_cells <- as.numeric(dt$total_cells)
dt$Proportion <- dt$N/dt$total_cells

p <- ggplot(data=dt, aes(x=Group, y=Proportion, fill=Major_T_types))
p <- p + geom_col(position="stack", color="black")
p <- p + theme_sara_90() + facet_wrap(~Group, nrow=1, scale="free_x")
p <- p + scale_fill_manual("", values=tcell.cols)
p <- p + labs(y="Proportion")
p <- p + theme(axis.title.x=element_blank(),
               axis.text=element_text(size=11),
               axis.text.x=element_blank(),
               legend.text=element_text(size=11),
               legend.title=element_text(size=1),
               axis.title.y=element_text(size=12),
               strip.text=element_text(size=11),
               legend.spacing.x = unit(0.5, 'cm'),
               legend.spacing.y=unit(0.2, "cm"),
               legend.key.size=unit(0.5, "cm"),
               legend.position="bottom") 
p <- p + guides(fill=guide_legend(byrow=TRUE))
# ggsave("./../processedData/Subclustering/Tcells_Bar_prop.png", width=10, height=6, dpi=400)
p

# pdf(file = "~/vincent_perez/Chang_T_cell_barplot.pdf",  
#     width = 6,
#     height = 5
#     ) 
# p
# dev.off()



```

```{r}

Idents(seurat_sub) <- seurat_sub@meta.data$Group
p1 <- DimPlot(seurat_sub, group.by="Major_T_types", cells=WhichCells(seurat_sub, idents="Control"))
p1 <- p1 + theme_sara() + labs(title="Control") + theme(legend.position="none")
p1 <- p1 + scale_color_manual(values=tcell.cols)

p2 <- DimPlot(seurat_sub, group.by="Major_T_types", cells=WhichCells(seurat_sub, idents="CpG"))
p2 <- p2 + theme_sara() + labs(title="CpG")  + theme(legend.position="none")
p2 <- p2 + scale_color_manual(values=tcell.cols)

p3 <- DimPlot(seurat_sub, group.by="Major_T_types", cells=WhichCells(seurat_sub, idents="RT"))
p3 <- p3 + theme_sara() + labs(title="RT")  + theme(legend.position="none")
p3 <- p3 + scale_color_manual(values=tcell.cols)

p4 <- DimPlot(seurat_sub, group.by="Major_T_types", cells=WhichCells(seurat_sub, idents="RT_plus_CpG"))
p4 <- p4 + theme_sara() + labs(title="RT_plus_CpG")  + theme(legend.position="none")
p4 <- p4 + scale_color_manual(values=tcell.cols)

p <- cowplot::plot_grid(p1, p2, p3, p4, ncol=2, nrow=2, labels="AUTO")
# ggsave("./../processedData/Subclustering/Tcells/Umap_MajorTypes.png", plot=p, width=12, height=10, units="in", device="png", dpi=400)
p

# pdf(file = "~/vincent_perez/Chang_T_cells_UMAP.pdf",  
#     width = 8,
#     height = 5
#     ) 
# cowplot::plot_grid(p1, p2, p3, p4, ncol=2, nrow=2, labels="AUTO")
# dev.off()
```


```{r}
p4 <- DimPlot(seurat_sub, group.by="Major_T_types")
p4 <- p4 + theme_sara()
p4 <- p4 + scale_color_manual(values=tcell.cols)
# ggsave("./../processedData/Subclustering/Tcells/Umap_MajorTypes_allsamples.png", plot=p4, width=6, height=4, units="in", device="png", dpi=400)
p4
```

# Subclustering - Method 2 - T cell predictors

```{r}
# ref <- ProjecTILs::load.reference.map()
```

```{r}
# tcell_filt.projected <- ProjecTILs::make.projection(seurat_sub, ref=ref, filter.cells=TRUE)
# tcell_filt.projected <- ProjecTILs::cellstate.predict(ref=ref, query=tcell_filt.projected)
# table(tcell_filt.projected$functional.cluster)
```


```{r}
# dat <- as.data.frame(seurat_sub@meta.data)
# dat <- dat[,names(dat) %in% c("SampleID", "tcell_projected_clust")]
# dat$rn <- rownames(seurat_sub@meta.data)
# tcell.df <- as.data.frame(tcell_filt.projected@meta.data)
# tcell.df <- tcell.df[, names(tcell.df) %in% c("functional.cluster", "SampleID")]
# tcell.df$rn <- rownames(tcell_filt.projected@meta.data)
# 
# meta <- merge(dat, tcell.df, all.x=TRUE, by="rn")
# perm <- match(rownames(seurat_sub@meta.data), meta$rn)
# meta <- meta[perm,]
# rownames(meta) <- meta$rn
# seurat_sub <- AddMetaData(seurat_sub, meta)
# seurat_sub@meta.data$tcell_projected_clust_filt <- ifelse(is.na(seurat_sub@meta.data$functional.cluster), "Other", seurat_sub@meta.data$functional.cluster)
```

## UMAPS with ProjectTILs
```{r, fig.width=12, fig.height=10}

# lcc <- c("#EA7580", "#F6A1A5", "#F8CD9C", "#1BB6AF", "#088BBE",
#          "#172869", "#C70E7B", "#FC6882", "#A6E000", "#D72000",
#          "#EE6100", "#FFAD0A")
# 
# 
# tcell.cols <- c(`CD4_NaiveLike`="#F6A1A5",
#                 `CD8_EarlyActiv`="#D72000",
#                 `CD8_EffectorMemory`="#172869",
#                 `CD8_NaiveLike`="#1BB6AF",
#                 `CD8_Tex`="#FFAD0A",
#                 `CD8_Tpex`="lightblue",
#                 `Other`="grey90" ,
#                 `Tfh`="#C70E7B",
#                 `Th1`="#088BBE",
#                 `Treg`="#A6E000")
# 
# p3 <- DimPlot(seurat_sub, group.by=subcluster.name)
# p3 <- p3 + theme_sara() + theme(legend.position = "bottom")
# p3 <- p3 + scale_color_manual(values=cols_cluster)
# 
# 
# cell_cycle.cols <- c("#01617E","#8B9934","#E04883")
# names(cell_cycle.cols) <- levels(as.factor(seurat_sub@meta.data$Phase))
# pt <- DimPlot(seurat_sub, group.by="Phase")
# pt <- pt + theme_sara() + theme(legend.position = "bottom")
# pt <- pt + scale_color_manual(values=cell_cycle.cols)
# 
# pt2 <- DimPlot(seurat_sub, group.by="tcell_projected_clust_filt")
# pt2 <- pt2 + theme_sara() + theme(legend.position = "bottom")
# pt2 <- pt2 + scale_color_manual(values=tcell.cols)
# 
# 
# immgen.cols <- ref.cols[["SingleR_ImmGen"]][which(names(ref.cols[["SingleR_ImmGen"]]) %in% params$cells)]
# pa <- DimPlot(seurat_sub, group.by="SingleR_ImmGen")
# pa <- pa + theme_sara() + theme(legend.position = "bottom")
# pa <- pa + scale_color_manual(values=immgen.cols)
# 
# p <- cowplot::plot_grid(p3, pa,
#                    pt, pt2, nrow=2, ncol=2, labels="AUTO", adjust="vh")
# # ggsave("./../processedData/Subclustering/Tcells/UMAP_all.png", width=12, height=10, units="in", device="png", dpi=400)
# 
# p
```

# UMAPs and STATs og GOIs

```{r}
### Add CD8a to the statistics
gois<-c("Gzmk", "Gzmb", "Cd8a","Ifng")
groups<-c(unique(seurat_sub@meta.data$Group2))
Idents(seurat_sub)<-seurat_sub@meta.data$Group2

seurats<-list("RT"=subset(seurat_sub, idents="RT"),
              "CpG+RT"=subset(seurat_sub, idents="CpG+RT"),
              "Ctl"=subset(seurat_sub, idents="Ctl"),
              "CpG"=subset(seurat_sub, idents="CpG"))
names(seurats)<-c("RT","CpG+RT","Ctl","CpG")

### GZMK
gzmk_plots<-list()
for (i in 1:length(seurats)) {
    p1 <- eval(substitute(

  FeaturePlot(seurats[[i]], features=gois[1],pt.size = 0.5) +
    ggtitle(paste0(names(seurats[i]), " - ",gois[1])) + 
    scale_color_viridis_c(option="rocket", direction=-1,limits = c(0, 5)) + 
    theme_sara()
            
      ,list(i = i)))
    gzmk_plots[[i]] <- p1  # add each plot into plot list
}

p1<-cowplot::plot_grid(plotlist = gzmk_plots, labels="AUTO", nrow=2, ncol=2)
p1

# ### Write to pdf
# pdf(file = "~/vincent_perez/figures/Chang_T_cell_GZMK_UMAP.pdf",
#     width = 8,
#     height = 6
#     )
# p1
# dev.off

#### GZMB
gzmb_plots<-list()
for (i in 1:length(seurats)) {
    p1 <- eval(substitute(

  FeaturePlot(seurats[[i]], features=gois[2],pt.size = 0.5) +
    ggtitle(paste0(names(seurats[i]), " - ",gois[2])) + 
    scale_color_viridis_c(option="rocket", direction=-1, limits=c(0,5)) + 
    theme_sara()
            
      ,list(i = i)))
    gzmb_plots[[i]] <- p1  # add each plot into plot list
}

p2<-cowplot::plot_grid(plotlist = gzmb_plots, labels="AUTO", nrow=2, ncol=2)
p2

# ### Write to PDF
# pdf(file = "~/vincent_perez/figures/Chang_T_cell_GZMB_UMAP.pdf",
#     width = 8,
#     height = 6
#     )
# p2
# dev.off

### CD8a
cd8a_plots<-list()
for (i in 1:length(seurats)) {
    p1 <- eval(substitute(

  FeaturePlot(seurats[[i]], features=gois[3],pt.size = 0.5) +
    ggtitle(paste0(names(seurats[i]), " - ",gois[3])) + 
    scale_color_viridis_c(option="rocket", direction=-1,limits=c(0,5)) + 
    theme_sara()
            
      ,list(i = i)))
    cd8a_plots[[i]] <- p1  # add each plot into plot list
}

p3<-cowplot::plot_grid(plotlist = cd8a_plots, labels="AUTO", nrow=2, ncol=2)
p3

# ### Write to pdf
# pdf(file = "~/vincent_perez/figures/Chang_T_cell_CD8a_UMAP.pdf",
#     width = 8,
#     height = 6
#     )
# p3
# dev.off

### IFNg
ifng_plots<-list()
for (i in 1:length(seurats)) {
    p1 <- eval(substitute(

  FeaturePlot(seurats[[i]], features=gois[4],pt.size = 0.5) +
    ggtitle(paste0(names(seurats[i]), " - ",gois[4])) + 
    scale_color_viridis_c(option="rocket", direction=-1,limits=c(0,5)) + 
    theme_sara()
            
      ,list(i = i)))
    ifng_plots[[i]] <- p1  # add each plot into plot list
}

p4<-cowplot::plot_grid(plotlist = ifng_plots, labels="AUTO", nrow=2, ncol=2)
p4

### Write to pdf
# pdf(file = "~/vincent_perez/figures/Chang_T_cell_IFNG_UMAP.pdf",
#     width = 8,
#     height = 6
#     )
# p4
# dev.off

### Run stats on the 4 genes
mycomparisons<-list(c("Ctl", "CpG+RT"),
                    c("CpG","CpG+RT"),
                    c("RT", "CpG+RT"))

### Run wilcoxon rank sum test for markers of interest
seurat_scc@meta.data$Gzmk<-as.vector(seurat_scc@assays$RNA@data[c("Gzmk"),])
seurat_scc@meta.data$Gzmb<-as.vector(seurat_scc@assays$RNA@data[c("Gzmb"),])
seurat_scc@meta.data$Cd8a<-as.vector(seurat_scc@assays$RNA@data[c("Cd8a"),])
seurat_scc@meta.data$Ifng<-as.vector(seurat_scc@assays$RNA@data[c("Ifng"),])
  
myplots <- list()  # new empty list
for (i in 1:length(gois)) {
    p1 <- eval(substitute(
      
ggplot(data=seurat_scc@meta.data,
       aes(x=Group2,
           y= get(gois[i]))) +
  geom_violin(aes(fill=Group2),
              width=1.0) +
  geom_boxplot(width=0.2, 
               color="black", 
               alpha=0.2,
               outlier.shape = NA) +
  scale_fill_manual("Group",values=group.cols2)+
  theme_sara()+
#  ylim(0.01, 8)+
  ggtitle(paste0(gois[i]))+
  stat_compare_means(comparisons = mycomparisons,
                     size=3
  ) +
  theme(legend.position = "none",
        axis.title.x = element_blank(),
        axis.title.y = element_blank())

      ,list(i = i)))
    myplots[[i]] <- p1  # add each plot into plot list
}

p<-cowplot::plot_grid(plotlist = myplots)
p

pdf(file = "~/vincent_perez/figures/Chang_T_cell_GZMB_GZMK_CD8a_IFNg_stats.pdf",
    width = 6,
    height = 6
    )
p
dev.off
```

```{r, fig.width=12, fig.height=10}
# p.list <- FeaturePlot(seurat_sub, features=c("Cd8a", "Tcf7", "Cd4", "Pdcd1"), combine=FALSE) 
# 
# for (i in 1:length(p.list)) {
#   p <- p.list[[i]]
#   p <- p + scale_color_viridis_c(option="rocket", direction=-1)
#   p <- p + theme_sara()
#   p.list[[i]] <- p
# }
# 
# p <- cowplot::plot_grid(plotlist = p.list, labels="AUTO", nrow=2, ncol=2)
# # ggsave("./../processedData/Subclustering/Tcells/UMAP_exhausted.png", width=12, height=10, units="in", device="png", dpi=400)
# 
# pp8 <- cowplot::plot_grid(p.list[[1]], labels="AUTO", nrow=1)
# pp4 <- cowplot::plot_grid(p.list[[3]], labels="AUTO", nrow=1)
# # ggsave("./../processedData/Subclustering/Tcells/UMAP_CD8.png", plot=pp8, width=6, height=4, units="in", device="png", dpi=400)
# # ggsave("./../processedData/Subclustering/Tcells/UMAP_CD4.png", plot=pp4, width=6, height=4, units="in", device="png", dpi=400)
# 
# p
```

# R Session Information
<details>
<summary>Information about R, the OS and attached or loaded packages</summary>
```{r sesion_info}
pander::pander(sessionInfo(), compact = FALSE)
```
</details>
***
<center>`r format(Sys.time(), '%d %B %Y')`</center>
